Re-organize integration test folder for the jobs
