+++ test/integration/controller/jobs/tfjob/suite_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package jobset
+package tfjob
 
 import (
 	"context"
@@ -33,7 +33,7 @@ import (
 	"sigs.k8s.io/kueue/pkg/controller/core"
 	"sigs.k8s.io/kueue/pkg/controller/core/indexer"
 	"sigs.k8s.io/kueue/pkg/controller/jobframework"
-	"sigs.k8s.io/kueue/pkg/controller/jobs/jobset"
+	"sigs.k8s.io/kueue/pkg/controller/jobs/kubeflow/jobs/tfjob"
 	"sigs.k8s.io/kueue/pkg/queue"
 	"sigs.k8s.io/kueue/pkg/scheduler"
 	"sigs.k8s.io/kueue/test/integration/framework"
@@ -41,34 +41,33 @@ import (
 )
 
 var (
-	cfg           *rest.Config
-	k8sClient     client.Client
-	ctx           context.Context
-	fwk           *framework.Framework
-	crdPath       = filepath.Join("..", "..", "..", "..", "config", "components", "crd", "bases")
-	jobsetCrdPath = filepath.Join("..", "..", "..", "..", "dep-crds", "jobset-operator")
-	webhookPath   = filepath.Join("..", "..", "..", "..", "config", "components", "webhook")
+	cfg               *rest.Config
+	k8sClient         client.Client
+	ctx               context.Context
+	fwk               *framework.Framework
+	crdPath           = filepath.Join("..", "..", "..", "..", "..", "config", "components", "crd", "bases")
+	tensorflowCrdPath = filepath.Join("..", "..", "..", "..", "..", "dep-crds", "training-operator")
 )
 
 func TestAPIs(t *testing.T) {
 	gomega.RegisterFailHandler(ginkgo.Fail)
 
 	ginkgo.RunSpecs(t,
-		"JobSet Controller Suite",
+		"TFJob Controller Suite",
 	)
 }
 
 func managerSetup(opts ...jobframework.Option) framework.ManagerSetup {
 	return func(mgr manager.Manager, ctx context.Context) {
-		reconciler := jobset.NewReconciler(
+		reconciler := tfjob.NewReconciler(
 			mgr.GetClient(),
 			mgr.GetEventRecorderFor(constants.JobControllerName),
 			opts...)
-		err := jobset.SetupIndexes(ctx, mgr.GetFieldIndexer())
+		err := tfjob.SetupIndexes(ctx, mgr.GetFieldIndexer())
 		gomega.Expect(err).NotTo(gomega.HaveOccurred())
 		err = reconciler.SetupWithManager(mgr)
 		gomega.Expect(err).NotTo(gomega.HaveOccurred())
-		err = jobset.SetupJobSetWebhook(mgr, opts...)
+		err = tfjob.SetupTFJobWebhook(mgr, opts...)
 		gomega.Expect(err).NotTo(gomega.HaveOccurred())
 	}
 }
@@ -84,12 +83,12 @@ func managerAndSchedulerSetup(opts ...jobframework.Option) framework.ManagerSetu
 		failedCtrl, err := core.SetupControllers(mgr, queues, cCache, &config.Configuration{})
 		gomega.Expect(err).ToNot(gomega.HaveOccurred(), "controller", failedCtrl)
 
-		err = jobset.SetupIndexes(ctx, mgr.GetFieldIndexer())
+		err = tfjob.SetupIndexes(ctx, mgr.GetFieldIndexer())
 		gomega.Expect(err).NotTo(gomega.HaveOccurred())
-		err = jobset.NewReconciler(mgr.GetClient(),
+		err = tfjob.NewReconciler(mgr.GetClient(),
 			mgr.GetEventRecorderFor(constants.JobControllerName), opts...).SetupWithManager(mgr)
 		gomega.Expect(err).NotTo(gomega.HaveOccurred())
-		err = jobset.SetupJobSetWebhook(mgr, opts...)
+		err = tfjob.SetupTFJobWebhook(mgr, opts...)
 		gomega.Expect(err).NotTo(gomega.HaveOccurred())
 
 		sched := scheduler.New(queues, cCache, mgr.GetClient(), mgr.GetEventRecorderFor(constants.AdmissionName))
